refactor: abstract packages (#8)
This PR abstracts the repo into 3 packages:

## Assets
The `@dying-sun/assets` package contains all game assets, as well as the
build system for packing images into spritesheets. This package should
be imported into other packages and will typically be mounted as the
`public` directory on a web server.

## Game
The `@dying-sun/game` package is just the game application. It's served
by Vite for easy development, but this package should not be deployed
alone. Instead, it should be imported into a hosting package, such as a
web server or an Electron app.

## Next Wrapper
The `@dying-sun/next-wrapper` package is a _very_ basic Next.js
application. It uses a symbolic link to serve the `@dying-sun/assets`
package as the `public` directory, and it only serves one page which
runs the `@dying-sun/game` package.

---------

Signed-off-by: Trezy <trezy@birb.house>